Create text runs during box building
