+++ src/components/main/servo.rc
@@ -70,13 +70,14 @@ pub mod layout {
     pub mod box_builder;
     pub mod context;
     pub mod display_list_builder;
+    pub mod float_context;
+    pub mod float;
     pub mod flow;
     pub mod layout_task;
     pub mod inline;
     pub mod model;
     pub mod text;
-    pub mod float_context;
-    pub mod float;
+    pub mod util;
     mod aux;
 }
